requirements.txt
setuptools==59.5.0
numpy==1.26.4
scipy==1.10.1
soundfile>=0.12.1
librosa==0.9.2
torch
torchaudio
speechbrain==0.5.16
pyannote-audio>=3.1.1
pyannote-whisper @ git+https://github.com/yinruiqing/pyannote-whisper.git
openai-whisper==20231117
pydub>=0.25.1
python-dotenv>=1.0.1
openai>=1.30.0
streamlit>=1.32